requirements.txt
numpy
scipy
matplotlib
h5py
tensorboard
lmdb
attrs
tqdm
jupyter
opencv-python
